backend/requirements.txt
# FastAPI Backend Dependencies
# AI-Trader Platform - 2025-10-29
# Compatible versions - no conflicts

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0

# Database
supabase>=2.0.0
psycopg2-binary>=2.9.0

# Data Processing (from aitrtader utils)
numpy>=1.24.0
pandas>=2.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Validation (FastAPI includes pydantic)
pydantic-settings>=2.0.0
email-validator>=2.0.0

# HTTP Client
httpx>=0.25.0

# Utilities
python-dateutil>=2.8.0

# AI Trading Engine (LangChain for AI agents)
langchain>=1.0.0
langchain-openai>=1.0.0
langchain-mcp-adapters>=0.1.0
fastmcp>=2.0.0

